fix: support multi-tenant mode
