+++ pkg/services/apiserver/wireset.go
@@ -10,8 +10,8 @@ var WireSet = wire.NewSet(
 	ProvideEventualRestConfigProvider,
 	wire.Bind(new(RestConfigProvider), new(*eventualRestConfigProvider)),
 	wire.Bind(new(DirectRestConfigProvider), new(*eventualRestConfigProvider)),
-	ProvideScheme,
-	ProvideCodecFactory,
+	builder.ProvideScheme,
+	builder.ProvideCodecFactory,
 	ProvideService,
 	wire.Bind(new(Service), new(*service)),
 	wire.Bind(new(builder.APIRegistrar), new(*service)),
